Team E/pyproject.toml
[project]
name = "muscle-music"
version = "0.1.0"
description = "Muscle Music Project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "numpy>=2.4.4",
    "pandas>=2.2.3",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "wandb>=0.25.1",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.nvidia.com"

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
